Add ansible galaxy role collection publishing
Move playbooks to the playbook directory.

Adjust .gitignore.

Add placeholder README.md files for each of the roles, this satisfies
the requirements for publishing to ansible galaxy.

Add a scripts directory that has a playbook for generating and
publishing the roles collection.

Update root README.md to capture steps for publishing and using
ansible-galaxy.

Change-Id: I10aa0f0f2b9664fd447f2863ef6217f593bcdfbc